+++ emmett/asgi/handlers.py
@@ -9,37 +9,49 @@
     :license: BSD-3-Clause
 """
 
+from __future__ import annotations
+
 import asyncio
+import concurrent
 import os
 import re
 
 from collections import OrderedDict
-from datetime import datetime
+from typing import Any, Awaitable, Callable, Optional, Tuple, Union
 
 from ..ctx import Context, current
 from ..debug import smart_traceback, debug_handler
 from ..http import HTTPResponse, HTTPFile, HTTP
 from ..language import T
 from ..utils import cachedprop
-from ..wrappers.request import Body, Request
+from ..wrappers.helpers import RequestCancelled
+from ..wrappers.request import Request
 from ..wrappers.response import Response
 from ..wrappers.websocket import Websocket
+from .typing import Event, EventHandler, EventLooper, Receive, Scope, Send
 
 REGEX_STATIC = re.compile(
-    '^/static/(?P<v>_\d+\.\d+\.\d+/)?(?P<f>.*?)$')
+    r'^/static/(?P<v>_\d+\.\d+\.\d+/)?(?P<f>.*?)$')
 REGEX_STATIC_LANG = re.compile(
-    '^/(?P<l>\w+/)?static/(?P<v>_\d+\.\d+\.\d+/)?(?P<f>.*?)$')
+    r'^/(?P<l>\w+/)?static/(?P<v>_\d+\.\d+\.\d+/)?(?P<f>.*?)$')
 
 
-class HandlerEvent:
+class EventHandlerWrapper:
     __slots__ = ['event', 'f']
 
-    def __init__(self, event, f):
+    def __init__(self, event: str, f: EventHandler):
         self.event = event
         self.f = f
 
-    async def __call__(self, *args, **kwargs):
-        task = await self.f(*args, **kwargs)
+    async def __call__(
+        self,
+        handler: Handler,
+        scope: Scope,
+        receive: Receive,
+        send: Send,
+        event: Event
+    ) -> Tuple[Optional[EventHandler], None]:
+        task = await self.f(handler, scope, receive, send, event)
         return task, None
 
 
@@ -50,7 +62,7 @@ class MetaHandler(type):
         all_events = OrderedDict()
         events = []
         for key, value in list(attrs.items()):
-            if isinstance(value, HandlerEvent):
+            if isinstance(value, EventHandlerWrapper):
                 events.append((key, value))
         declared_events.update(events)
         new_class._declared_events_ = declared_events
@@ -71,31 +83,60 @@ class Handler(metaclass=MetaHandler):
         self.app = app
 
     @classmethod
-    def on_event(cls, event):
-        def wrap(f):
-            return HandlerEvent(event, f)
+    def on_event(
+        cls, event: str
+    ) -> Callable[[EventHandler], EventHandlerWrapper]:
+        def wrap(f: EventHandler) -> EventHandlerWrapper:
+            return EventHandlerWrapper(event, f)
         return wrap
 
-    def get_event_handler(self, event_type):
-        return self._events_handlers_.get(event_type, _event_looper)
+    def get_event_handler(
+        self, event_type: str
+    ) -> Union[EventHandler, EventHandlerWrapper]:
+        return self._events_handlers_.get(event_type, _event_missing)
 
-    async def __call__(self, scope, receive, send):
-        await self.handle_events(scope, receive, send)
+    def __call__(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send
+    ) -> Awaitable[None]:
+        return self.handle_events(scope, receive, send)
 
-    async def handle_events(self, scope, receive, send):
-        task, event = _event_looper, None
+    async def handle_events(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send
+    ):
+        task: Optional[EventLooper] = _event_looper
+        event = None
         while task:
             task, event = await task(self, scope, receive, send, event)
 
 
 class LifeSpanHandler(Handler):
+    __slots__ = []
+
     @Handler.on_event('lifespan.startup')
-    async def event_startup(self, scope, receive, send, event):
+    async def event_startup(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send,
+        event: Event
+    ) -> EventLooper:
         await send({'type': 'lifespan.startup.complete'})
         return _event_looper
 
     @Handler.on_event('lifespan.shutdown')
-    async def event_shutdown(self, scope, receive, send, event):
+    async def event_shutdown(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send,
+        event: Event
+    ):
         await send({'type': 'lifespan.shutdown.complete'})
 
 
@@ -131,58 +172,59 @@ class HTTPHandler(RequestHandler):
             self._prefix_handler if self.router._prefix_main else
             self.static_handler)
 
-    async def __call__(self, scope, receive, send):
-        scope['emt.path'] = scope['path'] or '/'
-        scope['emt.now'] = datetime.utcnow()
-        scope['emt.input'] = Body(self.app.config.request_max_content_length)
-        task_events = asyncio.create_task(
-            self.handle_events(scope, receive, send))
-        task_request = asyncio.create_task(
-            self.handle_request(scope, receive, send))
-        _, pending = await asyncio.wait(
-            [task_request, task_events], return_when=asyncio.FIRST_COMPLETED
-        )
-        await _cancel_tasks(pending)
-
-    @Handler.on_event('http.disconnect')
-    async def event_disconnect(self, scope, receive, send, event):
-        return
-
-    @Handler.on_event('http.request')
-    async def event_request(self, scope, receive, send, event):
-        scope['emt.input'].append(event['body'])
-        if not event.get('more_body', False):
-            scope['emt.input'].set_complete()
-        return _event_looper
+    def __call__(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send
+    ) -> Awaitable[None]:
+        return self.handle_request(scope, receive, send)
 
     @cachedprop
-    def error_handler(self):
+    def error_handler(self) -> Callable[[], Awaitable[str]]:
         return (
-            self._debug_handler if self.app.debug else self.exception_handler)
+            self._debug_handler if self.app.debug else self.exception_handler
+        )
 
     @cachedprop
-    def exception_handler(self):
+    def exception_handler(self) -> Callable[[], Awaitable[str]]:
         return self.app.error_handlers.get(500, self._exception_handler)
 
-    async def handle_request(self, scope, receive, send):
+    async def handle_request(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send
+    ):
+        scope['emt.path'] = scope['path'] or '/'
         try:
-            http = await asyncio.shield(self.pre_handler(scope, receive, send))
-        except HTTPResponse as httpe:
-            http = httpe
-        except asyncio.CancelledError:
+            http = await self.pre_handler(scope, receive, send)
+        except RequestCancelled:
             return
-        # TODO: timeout from app config/response
-        await asyncio.wait_for(http.send(scope, send), None)
+        await asyncio.wait_for(
+            http.send(scope, send),
+            self.app.config.response_timeout
+        )
 
-    def _prefix_handler(self, scope, receive, send):
-        path = current.request.path
+    @staticmethod
+    async def _http_response(code: int) -> HTTPResponse:
+        return HTTP(code)
+
+    def _prefix_handler(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send
+    ) -> Awaitable[HTTPResponse]:
+        path = scope['emt.path']
         if not path.startswith(self.router._prefix_main):
-            return HTTP(404)
-        current.request.path = scope['emt.path'] = (
-            path[self.router._prefix_main_len:] or '/')
+            return self._http_response(404)
+        scope['emt.path'] = path[self.router._prefix_main_len:] or '/'
         return self.static_handler(scope, receive, send)
 
-    def _static_lang_matcher(self, path):
+    def _static_lang_matcher(
+        self, path: str
+    ) -> Tuple[Optional[str], Optional[str]]:
         match = REGEX_STATIC_LANG.match(path)
         if match:
             lang, version, file_name = match.group('l', 'v', 'f')
@@ -194,17 +236,24 @@ class HTTPHandler(RequestHandler):
             return static_file, version
         return None, None
 
-    def _static_nolang_matcher(self, path):
+    def _static_nolang_matcher(
+        self, path: str
+    ) -> Tuple[Optional[str], Optional[str]]:
         if path.startswith('/static'):
             version, file_name = REGEX_STATIC.match(path).group('v', 'f')
             static_file = os.path.join(self.app.static_path, file_name)
             return static_file, version
         return None, None
 
-    async def _static_response(self, file_path):
+    async def _static_response(self, file_path: str) -> HTTPFile:
         return HTTPFile(file_path)
 
-    def _static_handler(self, scope, receive, send):
+    def _static_handler(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send
+    ) -> Awaitable[HTTPResponse]:
         path = scope['emt.path']
         #: handle internal assets
         if path.startswith('/__emmett__'):
@@ -212,17 +261,24 @@ class HTTPHandler(RequestHandler):
             static_file = os.path.join(
                 os.path.dirname(__file__), '..', 'assets', file_name)
             if os.path.splitext(static_file)[1] == 'html':
-                return HTTP(404)
+                return self._http_response(404)
             return self._static_response(static_file)
         #: handle app assets
-        static_file, version = self.static_matcher(path)
+        static_file, _ = self.static_matcher(path)
         if static_file:
             return self._static_response(static_file)
         return self.dynamic_handler(scope, receive, send)
 
-    async def dynamic_handler(self, scope, receive, send):
+    async def dynamic_handler(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send
+    ) -> HTTPResponse:
+        ctx_token = current._init_(
+            RequestContext, self.app, scope, receive, send
+        )
         try:
-            ctx_token = current._init_(RequestContext, self.app, scope)
             http = await self.router.dispatch()
         except HTTPResponse as http_exception:
             http = http_exception
@@ -236,9 +292,11 @@ class HTTPHandler(RequestHandler):
                 )
             #: always set cookies
             http.set_cookies(current.response.cookies)
+        except RequestCancelled:
+            raise
         except Exception:
             self.app.log.exception('Application exception:')
-            raise HTTP(
+            http = HTTP(
                 500,
                 await self.error_handler(),
                 headers=current.response.headers
@@ -247,10 +305,10 @@ class HTTPHandler(RequestHandler):
             current._close_(ctx_token)
         return http
 
-    async def _debug_handler(self):
+    async def _debug_handler(self) -> str:
         return debug_handler(smart_traceback(self.app))
 
-    async def _exception_handler(self):
+    async def _exception_handler(self) -> str:
         current.response.headers['Content-Type'] = 'text/plain'
         return 'Internal error'
 
@@ -266,8 +324,12 @@ class WSHandler(RequestHandler):
             self._prefix_handler if self.router._prefix_main else
             self.dynamic_handler)
 
-    async def __call__(self, scope, receive, send):
-        scope['emt.path'] = scope['path'] or '/'
+    async def __call__(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send
+    ):
         scope['emt.input'] = asyncio.Queue()
         task_events = asyncio.create_task(
             self.handle_events(scope, receive, send))
@@ -277,51 +339,100 @@ class WSHandler(RequestHandler):
             [task_request, task_events], return_when=asyncio.FIRST_COMPLETED
         )
         scope['emt._flow_cancel'] = True
-        await _cancel_tasks(pending)
+        _cancel_tasks(pending)
+
+    @Handler.on_event('websocket.connect')
+    async def event_connect(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send,
+        event: Event
+    ) -> EventLooper:
+        return _event_looper
 
     @Handler.on_event('websocket.disconnect')
-    async def event_disconnect(self, scope, receive, send, event):
+    async def event_disconnect(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send,
+        event: Event
+    ):
         return
 
     @Handler.on_event('websocket.receive')
-    async def event_receive(self, scope, receive, send, event):
+    async def event_receive(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send,
+        event: Event
+    ) -> EventLooper:
         await scope['emt.input'].put(event.get('bytes') or event['text'])
         return _event_looper
 
-    async def handle_request(self, scope, receive, send):
+    async def handle_request(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send
+    ):
+        scope['emt.path'] = scope['path'] or '/'
+        scope['emt._ws_closed'] = False
         try:
             await self.pre_handler(scope, receive, send)
+        except HTTPResponse:
+            if not scope['emt._ws_closed']:
+                await send({'type': 'websocket.close', 'code': 1006})
         except asyncio.CancelledError:
             if not scope.get('emt._flow_cancel', False):
                 self.app.log.exception('Application exception:')
         except Exception:
+            if not scope['emt._ws_closed']:
+                await send({'type': 'websocket.close', 'code': 1006})
             self.app.log.exception('Application exception:')
 
-    def _prefix_handler(self, scope, receive, send):
-        path = current.request.path
+    def _prefix_handler(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send
+    ) -> Awaitable[None]:
+        path = scope['emt.path']
         if not path.startswith(self.router._prefix_main):
-            return HTTP(404)
-        current.request.path = scope['emt.path'] = (
-            path[self.router._prefix_main_len:] or '/')
+            raise HTTP(404)
+        scope['emt.path'] = path[self.router._prefix_main_len:] or '/'
         return self.dynamic_handler(scope, receive, send)
 
-    async def dynamic_handler(self, scope, receive, send):
+    async def dynamic_handler(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send
+    ):
         ctx_token = current._init_(
             WSContext, self.app, scope, scope['emt.input'].get, send
         )
         try:
             await self.router.dispatch()
         finally:
-            if current.websocket._accepted:
+            if (
+                not scope.get('emt._flow_cancel', False) and
+                current.websocket._accepted
+            ):
                 await send({'type': 'websocket.close', 'code': 1000})
+                scope['emt._ws_closed'] = True
             current._close_(ctx_token)
 
 
 class RequestContext:
-    def __init__(self, app, scope):
+    def __init__(self, app, scope, receive, send):
         self.app = app
         self.request = Request(
             scope,
+            receive,
+            send,
             app.config.request_max_content_length,
             app.config.request_body_timeout
         )
@@ -352,17 +463,28 @@ class WSContext(Context):
         return self.websocket.accept_language.best_match(list(T._langmap))
 
 
-async def _event_looper(handler, scope, receive, send, event):
+async def _event_looper(
+    handler: Handler,
+    scope: Scope,
+    receive: Receive,
+    send: Send,
+    event: Any = None
+) -> Tuple[Union[EventHandler, EventHandlerWrapper], Event]:
     event = await receive()
     event_handler = handler.get_event_handler(event['type'])
     return event_handler, event
 
 
-async def _event_missing(handler, receive, send):
-    raise RuntimeError('Event type not recognized.')
+async def _event_missing(
+    handler: Handler,
+    scope: Scope,
+    receive: Receive,
+    send: Send,
+    event: Event
+):
+    raise RuntimeError(f"Event type '{event['type']}' not recognized")
 
 
-async def _cancel_tasks(tasks):
+def _cancel_tasks(tasks):
     for task in tasks:
         task.cancel()
-    await asyncio.gather(*tasks, return_exceptions=True)