+++ emmett/asgi/workers.py
@@ -9,11 +9,79 @@
     :license: BSD-3-Clause
 """
 
-from uvicorn.workers import UvicornWorker as _UvicornWorker
+import asyncio
+import logging
+
+from gunicorn.workers.base import Worker as _Worker
+
+from .loops import loops
+from .protocols import protocols_http, protocols_ws
+from .server import Config, Server
 
 
-class EmmettWorker(_UvicornWorker):
-    CONFIG_KWARGS = {
+class Worker(_Worker):
+    EMMETT_CONFIG = {}
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        logger = logging.getLogger("uvicorn.error")
+        logger.handlers = self.log.error_log.handlers
+        logger.setLevel(self.log.error_log.level)
+
+        logger = logging.getLogger("uvicorn.access")
+        logger.handlers = self.log.access_log.handlers
+        logger.setLevel(self.log.access_log.level)
+
+        config = {
+            "app": None,
+            "log_config": None,
+            "timeout_keep_alive": self.cfg.keepalive,
+            "timeout_notify": self.timeout,
+            "callback_notify": self.callback_notify,
+            "limit_max_requests": self.max_requests
+        }
+
+        if self.cfg.is_ssl:
+            config.update(
+                ssl_keyfile=self.cfg.ssl_options.get("keyfile"),
+                ssl_certfile=self.cfg.ssl_options.get("certfile"),
+                ssl_version=self.cfg.ssl_options.get("ssl_version"),
+                ssl_cert_reqs=self.cfg.ssl_options.get("cert_reqs"),
+                ssl_ca_certs=self.cfg.ssl_options.get("ca_certs"),
+                ssl_ciphers=self.cfg.ssl_options.get("ciphers")
+            )
+
+        if self.cfg.settings["backlog"].value:
+            config["backlog"] = self.cfg.settings["backlog"].value
+
+        config.update(self.EMMETT_CONFIG)
+        config.update(
+            http=protocols_http.get_protocol(config.get('http', 'auto')),
+            ws=protocols_ws.get_protocol(config.get('ws', 'auto'))
+        )
+
+        self.config = Config(**config)
+
+    def init_process(self):
+        self.config.loop = loops.get_loop(self.config.loop)
+        super().init_process()
+
+    def init_signals(self):
+        return
+
+    def run(self):
+        self.config.app = self.wsgi
+        server = Server(config=self.config)
+        loop = asyncio.get_event_loop()
+        loop.run_until_complete(server.serve(sockets=self.sockets))
+
+    async def callback_notify(self):
+        self.notify()
+
+
+class EmmettWorker(Worker):
+    EMMETT_CONFIG = {
         "loop": "uvloop",
         "http": "httptools",
         "proxy_headers": False,
@@ -22,7 +90,7 @@ class EmmettWorker(_UvicornWorker):
 
 
 class EmmettH11Worker(EmmettWorker):
-    CONFIG_KWARGS = {
+    EMMETT_CONFIG = {
         "loop": "asyncio",
         "http": "h11",
         "proxy_headers": False,