+++ emmett/wrappers/websocket.py
@@ -9,19 +9,22 @@
     :license: BSD-3-Clause
 """
 
+from typing import Dict, List, Optional, Tuple
+
+from ..asgi.typing import Scope, Receive, Send
 from . import ScopeWrapper
 
 
 class Websocket(ScopeWrapper):
-    __slots__ = [
-        '_wreceive', '_wsend', '_accepted', 'receive', 'send',
-        '_flow_receive', '_flow_send'
-    ]
+    __slots__ = ('_accepted', 'receive', 'send', '_flow_receive', '_flow_send')
 
-    def __init__(self, scope, receive, send):
-        super().__init__(scope)
-        self._wreceive = receive
-        self._wsend = send
+    def __init__(
+        self,
+        scope: Scope,
+        receive: Receive,
+        send: Send
+    ):
+        super().__init__(scope, receive, send)
         self._accepted = False
         self._flow_receive = None
         self._flow_send = None
@@ -32,15 +35,35 @@ class Websocket(ScopeWrapper):
         self._flow_receive = flow_receive
         self._flow_send = flow_send
 
-    async def accept(self):
+    @property
+    def _asgi_spec_version(self) -> int:
+        return int(''.join(
+            self._scope.get('asgi', {}).get('spec_version', '2.0').split('.')
+        ))
+
+    def _encode_headers(
+        self,
+        headers: Dict[str, str]
+    ) -> List[Tuple[bytes, bytes]]:
+        return [
+            (key.encode('utf-8'), val.encode('utf-8'))
+            for key, val in headers.items()
+        ]
+
+    async def accept(
+        self,
+        headers: Optional[Dict[str, str]] = None,
+        subprotocol: Optional[str] = None
+    ):
         if self._accepted:
             return
-        message = {'type': 'websocket.accept', 'subprotocol': None}
-        # TODO headers
-        await self._wsend(message)
+        message = {'type': 'websocket.accept', 'subprotocol': subprotocol}
+        if headers and self._asgi_spec_version > 20:
+            message['headers'] = self._encode_headers(headers)
+        await self._send(message)
         self._accepted = True
-        self.receive = self._receive
-        self.send = self._send
+        self.receive = self._wrapped_receive
+        self.send = self._wrapped_send
 
     async def _accept_and_receive(self):
         await self.accept()
@@ -50,16 +73,16 @@ class Websocket(ScopeWrapper):
         await self.accept()
         await self.send(data)
 
-    async def _receive(self):
-        data = await self._wreceive()
+    async def _wrapped_receive(self):
+        data = await self._receive()
         for method in self._flow_receive:
             data = method(data)
         return data
 
-    async def _send(self, data):
+    async def _wrapped_send(self, data):
         for method in self._flow_send:
             data = method(data)
         if isinstance(data, str):
-            await self._wsend({'type': 'websocket.send', 'text': data})
+            await self._send({'type': 'websocket.send', 'text': data})
         else:
-            await self._wsend({'type': 'websocket.send', 'bytes': data})
+            await self._send({'type': 'websocket.send', 'bytes': data})