+++ tests/helpers.py
@@ -20,15 +20,15 @@ json_dump = Serializers.get_for('json')
 
 
 class FakeRequestContext(RequestContext):
-    def __init__(self, app, scope):
+    def __init__(self, app, scope, *args):
         self.app = app
-        self.request = Request(scope)
+        self.request = Request(scope, None, None)
         self.response = Response()
         self.session = None
 
 
 class FakeWSContext(WSContext):
-    def __init__(self, app, scope):
+    def __init__(self, app, scope, *args):
         self.app = app
         self.websocket = Websocket(
             scope,
@@ -48,7 +48,9 @@ class FakeWSContext(WSContext):
 @contextmanager
 def current_ctx(path, app=None):
     builder = ScopeBuilder(path)
-    token = current._init_(FakeRequestContext, app, builder.get_data()[0])
+    token = current._init_(
+        FakeRequestContext, app, builder.get_data()[0], None, None
+    )
     yield current
     current._close_(token)
 
@@ -58,6 +60,6 @@ def ws_ctx(path, app=None):
     builder = ScopeBuilder(path)
     scope_data = builder.get_data()[0]
     scope_data.update(type='websocket', scheme='wss')
-    token = current._init_(FakeWSContext, app, scope_data)
+    token = current._init_(FakeWSContext, app, scope_data, None, None)
     yield current
     current._close_(token)